Forward background queue exceptions on main thread
Currently, when a crash occurs on a dispatch queue other than the main queue on mobile (iOS and Android), it is "silent", meaning that it causes that "background thread" to crash without impacting the UI thread.

This lead to weird behaviors in our app (for computing that relied on that queue to execute) without any way to that an exception was raised.

With these changes, we add a new configuration flag to enable exceptions forwarding on the main thread so that the app actually crashes and stops its execution. This new flag is enabled by default, but can be disabled.

(Opinion) It is not recommended to disable this configuration flag as your apps might crash in production without any obvious signs. You should always be aware of a raised exception and handle it accordingly in your app.

(Note) As the concept of "main queue" / "ui thread" is non existent in a jvmMain environment, a new actual implementation of OperationDispatchQueue has been created for androidMain (ExecutorDispatchQueue), and JvmDispatchQueue has been moved from jvmShared to jvmMain.